Allow to disable logging errors to file
In some cases, for example when capturing stdout output, logging errors to
a file becomes redundant.

Add a new `SOLARUS_FILE_LOGGING` CMake variable that, when set to `ON`,
passes a compiler define to enable logging errors to a file in the
Logger component. This variable is set to `ON` by default to keep
current behaviour.

Also make documentation of Debug functions more clear.