Fix restoring from fullscreen mode in Windows
In Windows, when returning from fullscreen to windowed mode, the
video output is still being rendered at fullscreen size inside the
restored window. The reason is that `on_window_resized()` is not
called from `set_window_size()` because the restored window size
and the previously stored window size are the same.

The solution is simply to call `on_window_resized()` directly.