Do not force OpenAL audio context to 32 Khz
It is not really needed and some audio hardware might not even support
this frequency, falling back anyway to other freqs and producing an
annoying error/warning (if OpenAL is set verbose), for example:

    AL lib: (EE) UpdateDeviceParams: Failed to set 32000hz, got 48000hz instead

Moreover, the 32 Khz sample rate required for SPC is already configured
in the respective buffer in `src/audio/Music.cpp`:

    alBufferData(destination_buffer, AL_FORMAT_STEREO16, raw_data.data(), nb_samples * 2, 32000);

Furthermore, for hardware that does support 32 Khz, forcing to this frequency
limits other buffers such as Ogg and IT data that use higher sample rates,
e.g. 44.1 Khz, where they instead get down-sampled to 32 Khz.