Fix random crashes with coroutines unit test
Random movement is not reliable for the coroutines unit test.

If the test (1269) is repeated enough times, it fails:

    [Solarus] [7400] Fatal: In timer callback:
        maps/bugs/1269_coroutines.lua:39:
        'on_changed' has not been called after 5000 [ms]

As suggested by @stdgregwar in Discord, we can fix the test by
creating a straight movement instead and change it with a
timer after starting the movement.

The fixing code was kindly provided by @llamazing.